fix: isolate solana raw fallback tests
